requirements.txt
appdirs==1.4.4
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
ci-info==0.2.0
click==8.0.3
configobj==5.0.6
configparser==5.2.0
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.2.5
etelemetry==0.3.0
fake-useragent==0.1.11
filelock==3.7.1
Flask==2.0.2
future==0.18.2
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.10.1
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
looseversion==1.0.1
lxml==4.7.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
networkx==2.6.3
nibabel==4.0.1
nipype==1.8.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parse==1.19.0
pathlib==1.0.1
Pillow==9.0.1
protobuf==3.17.3
prov==2.0.0
psycopg2-binary==2.9.3
pydot==1.4.2
pyee==8.2.2
PyMuPDF==1.19.5
pyparsing==3.0.9
pypng==0.0.21
pyppeteer==1.0.2
PyQRCode==1.2.1
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.1
pyxnat==1.4
qrcode==7.3.1
rdflib==6.2.0
requests==2.27.1
scipy==1.7.3
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
tqdm==4.62.3
traits==6.3.2
typing-extensions==3.10.0.0
urllib3==1.26.8
w3lib==1.22.0
websockets==10.1
Werkzeug==2.0.3
zipp==3.7.0
django-storages
boto3
